Load planets data on the startup
In order for the frontend to have all the planets data, the planets
model must be built in a way that it return all the planet data
on the startup and not just returns an empty array of planets.

Import the kepler-data and the logic of finding habitable planets from
the "Planets Project" into the planet model.

Add csv-parse package for the planet model code to stream files
correctly and return the right data.

Modify the code that it returns a promise and awaits till all the
planets have been stored on resolved, and then exports.

Use this functionality in server.js so that that planets data is
available on the startup.