Implement similar mcv pattern in launches & planets
For both the launches and planets model, Implement similar mvc
pattern where the controller is only responsible for working with
req and res and doesn't have to worry about the how the data is
being read or written to the model.

Extract out all of the implementation detail of getting the data in
appropriate format within in the model only.