Implete abortLaunch feature
In server, create an endpoint "/launches/:id" that the frontend can
talk to abort/delete the particular launch provided a launch id.

Build appropriate helper functions in launches model and controller
that allow to do so.

In the client, use this api and add the logic that aborts that
particular launch by id from the upcoming page.